Added token authorization for deleting users or items in admin panel

+++ backend/controllers/users.js
@@ -86,6 +86,12 @@ usersRouter.put('/:id', async (request, response) => {
 
 usersRouter.delete('/:id', async (request, response) => {
 
+  const token = getToken(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
   await User.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
